+++ usbhid_descriptor.h
@@ -176,7 +176,7 @@ typedef struct __packed {
 
 mbed_error_t      usbhid_get_descriptor(uint8_t             iface_id,
                                         uint8_t            *buf,
-                                        uint32_t           *desc_size,
+                                        uint8_t            *desc_size,
                                         uint32_t            usbdci_handler __attribute__((unused)));
